implement logout and access control

+++ base/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render, redirect
 from django.contrib.auth.models import User
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from .models import Room, Topic
 from .forms import RoomForm
@@ -9,6 +10,8 @@ from .forms import RoomForm
 # Create your views here.
 
 def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect('home')
 
     if request.method == "POST":
         username = request.POST.get('username')
@@ -27,6 +30,10 @@ def loginPage(request):
 
     context = {}
     return render(request,'base/login_register.html',context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -49,6 +56,7 @@ def room(request,pk):
     context = {'room' : selectedRoom}
     return render(request,'base/room.html',context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == "POST":
@@ -59,9 +67,14 @@ def createRoom(request):
     context  = {'form':form}
     return render(request,'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        messages.error(request,'Bad Request')
+        return redirect('home')
 
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
@@ -72,8 +85,14 @@ def updateRoom(request,pk):
     context = {'form':form}
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        messages.error(request,'Bad Request')
+        return redirect('home')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
